projects/18-gpu-accelerated-computing/requirements.txt
# GPU computing
cupy-cuda12x>=12.0

# Numerical computing
numpy>=1.25
scipy>=1.11

# Distributed computing
dask[distributed]>=2024.1

# Monitoring
prometheus-client>=0.18

# Utilities
pyyaml>=6.0
